Session: renaming module name for clarity
