+++ mlte/session/__init__.py
@@ -1,3 +1,3 @@
-from .state import Session, session, set_context, set_store
+from .session import Session, session, set_context, set_store
 
 __all__ = ["session", "Session", "set_context", "set_store"]
