+++ test/artifact/test_artifact.py
@@ -13,7 +13,7 @@ from mlte.context.context import Context
 from mlte.evidence.metadata import EvidenceMetadata, Identifier
 from mlte.negotiation.artifact import NegotiationCard
 from mlte.report.artifact import Report
-from mlte.session.state import set_context, set_store
+from mlte.session.session import set_context, set_store
 from mlte.spec.spec import Spec
 from mlte.store.artifact.store import ArtifactStore
 from mlte.store.base import StoreType, StoreURI
